requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.7
python-dotenv==0.19.0
boto3==1.26.0
Werkzeug==2.0.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-flask==1.2.0
requests==2.31.0
PyJWT==2.8.0
passlib==1.7.4
Flask-CORS==4.0.0 